+++ ChangeLog.md
@@ -1,5 +1,27 @@
 # Changelog for dom-events
 
+## 0.0.0.5
+
+- Creates interfaces for `Event`, `UIEvent` and `MouseEvent`.
+
+## 0.0.0.4
+
+- Adds `target` property to all events.
+
+## 0.0.0.3
+
+- Changes `which` to an `Int`
+
+## 0.0.0.2
+
+- Adds `stopPropagation`
+- Removes `createEvent`
+
+## 0.0.0.1
+
+- Adds `MouseEvent`
+- Fixes typo in `PointerEvent`
+
 ## 0.0.0.0
 
 - Initial release